* use predictable, depentable types, use 64bit intermediate types 	  for the 16bit iterators, as the formerly used 32bit types 	  overflow in some algorithms, ...

+++ trunk/lib/ImageIterator2.hh
@@ -28,7 +28,7 @@ public:
   class accu
   {
   public:
-    typedef signed int vtype;
+    typedef int32_t vtype;
     vtype v1, v2, v3;
     
     accu () { v1 = v2 = v3 = 0; }
@@ -167,7 +167,7 @@ public:
   class accu
   {
   public:
-    typedef signed int vtype;
+    typedef int32_t vtype;
     vtype v1, v2, v3, v4;
     
     accu () { v1 = v2 = v3 = v4 = 0; }
@@ -316,7 +316,7 @@ public:
   class accu
   {
   public:
-    typedef signed int vtype;
+    typedef int64_t vtype;
     vtype v1, v2, v3;
     
     accu () { v1 = v2 = v3 = 0; }
@@ -455,7 +455,7 @@ public:
   class accu
   {
   public:
-    typedef signed int vtype;
+    typedef int32_t vtype;
     vtype v1;
     
     accu () { v1 = 0; }
@@ -573,7 +573,7 @@ public:
   class accu
   {
   public:
-    typedef signed int vtype;
+    typedef int64_t vtype;
     vtype v1;
     
     accu () { v1 = 0; }
